# services/api-gateway/docs/swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
  httpx.ErrorBody:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      message:
        example: invalid request
        type: string
    type: object
  httpx.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httpx.ErrorBody'
    type: object
...
  main.listStockResponse:
    properties:
      count:
        type: integer
      items:
        items:
          $ref: '#/definitions/main.stockItem'
        type: array
    type: object
...
  main.otpRequestBody:
    properties:
      phone:
        type: string
    type: object
  main.otpRequestResponse:
    properties:
      dev_code:
        type: string
      expires_in_sec:
        type: integer
      ok:
        type: boolean
      phone_masked:
        type: string
      resend_after_sec:
        type: integer
    type: object
...
  main.postMovementBody:
    properties:
      delta:
        description: 'ADJUST: required, != 0'
        type: integer
      movement_type:
        type: string
      name:
        description: required when creating stock on first IN
        type: string
      note:
        type: string
      product_id:
        type: string
      qty:
        description: 'IN/OUT: required, > 0'
        type: integer
      ref_id:
        type: string
      ref_type:
        description: default MANUAL
        type: string
      reorder_level:
        description: optional on first IN
        type: integer
      sku:
        description: required when creating stock on first IN
        type: string
      unit_cost:
        description: 'IN: required >= 0; ADJUST: optional cost update'
        type: integer
    type: object
  main.postMovementResponse:
    properties:
      item:
        $ref: '#/definitions/main.stockItem'
      movement:
        $ref: '#/definitions/main.stockMovement'
    type: object
...
  main.stockItem:
    properties:
      cost_price:
        type: integer
      name:
        type: string
      on_hand:
        type: integer
      product_id:
        type: string
      reorder_level:
        type: integer
      sku:
        type: string
      updated_at:
        type: string
    type: object
...
  main.stockMovement:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      delta:
        description: signed effect on on_hand
        type: integer
      id:
        type: string
      movement_type:
        type: string
      note:
        type: string
      product_id:
        type: string
      qty:
        type: integer
      ref_id:
        type: string
      ref_type:
        type: string
      unit_cost:
        type: integer
    type: object
...
info:
  contact: {}
  description: Public API gateway for Gas Tam De customers and administrators.
  title: Gas Tam De API
  version: "1.0"
paths:
...
  /admin/inventory:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.listStockResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List inventory
      tags:
      - Admin - Inventory
    post:
      consumes:
      - application/json
      description: Supports IN, OUT, and ADJUST movements.
      parameters:
      - description: Inventory movement
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.postMovementBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.postMovementResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Record an inventory movement
      tags:
      - Admin - Inventory
...
  /auth/otp/request:
    post:
      consumes:
      - application/json
      parameters:
      - description: Phone number
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.otpRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.otpRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Request an OTP
      tags:
      - Authentication
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the access token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"